[Refactor] update init path
