+++ rl4co/envs/routing/__init__.py
@@ -1,12 +1,12 @@
-from rl4co.envs.routing.atsp import ATSPEnv
-from rl4co.envs.routing.cvrp import CVRPEnv
-from rl4co.envs.routing.cvrptw import CVRPTWEnv
-from rl4co.envs.routing.mtsp import MTSPEnv
-from rl4co.envs.routing.op import OPEnv
-from rl4co.envs.routing.pctsp import PCTSPEnv
-from rl4co.envs.routing.pdp import PDPEnv
-from rl4co.envs.routing.sdvrp import SDVRPEnv
-from rl4co.envs.routing.spctsp import SPCTSPEnv
-from rl4co.envs.routing.svrp import SVRPEnv
-from rl4co.envs.routing.tsp import TSPEnv
-from rl4co.envs.routing.mdcpdp import MDCPDPEnv
+from rl4co.envs.routing.atsp.env import ATSPEnv
+from rl4co.envs.routing.cvrp.env import CVRPEnv
+from rl4co.envs.routing.cvrptw.env import CVRPTWEnv
+from rl4co.envs.routing.mtsp.env import MTSPEnv
+from rl4co.envs.routing.op.env import OPEnv
+from rl4co.envs.routing.pctsp.env import PCTSPEnv
+from rl4co.envs.routing.pdp.env import PDPEnv
+from rl4co.envs.routing.sdvrp.env import SDVRPEnv
+from rl4co.envs.routing.spctsp.env import SPCTSPEnv
+from rl4co.envs.routing.svrp.env import SVRPEnv
+from rl4co.envs.routing.tsp.env import TSPEnv
+from rl4co.envs.routing.mdcpdp.env import MDCPDPEnv
