+++ rl4co/envs/scheduling/__init__.py
@@ -1,2 +1,2 @@
-from rl4co.envs.scheduling.ffsp import FFSPEnv
-from rl4co.envs.scheduling.smtwtp import SMTWTPEnv
+from rl4co.envs.scheduling.ffsp.env import FFSPEnv
+from rl4co.envs.scheduling.smtwtp.env import SMTWTPEnv
